Small refactor to move preparing the workspace as a concern of the workspace factory (for reuse later)
